requirements.txt
django
psycopg2
djangorestframework
django-environ
os
lxml